==> 01_unique_asins.ipynb <==
import pandas as pd


harvard_df = pd.read_csv("./data/harvard/amazon-purchases.csv")
harvard_asins = harvard_df["asin"].drop_duplicates()

# Save the unique ASINs to a new pkl file
harvard_asins.to_pickle("./data/pfw/01_unique_harvard_asins.pkl")

print("Unique ASINs have been written to '01_unique_harvard_asins.pkl'.")

ucsd_df = pd.read_csv('./data/ucsd/Books.csv')
ucsd_asins = ucsd_df['asin'].drop_duplicates()

# Save the unique ASINs to a new pkl file
ucsd_asins.to_pickle('./data/pfw/01_unique_ucsd_asins.pkl')
print("Unique ASINs have been written to '01_unique_ucsd_asins.pkl'.")
