==> 03_create_sample_asins.ipynb <==
# Create CSV file for Review data


import pandas as pd

sample_size = 100000
# Load data set
asins_pkl = pd.read_pickle("./data/pfw/02_combined_unique_asins_final.pkl")
sample_set = asins_pkl["asin"].sample(sample_size)
# sample_set.to_csv('./data/pfw/03_sample_asins.csv', index=False)
sample_set.to_pickle("./data/pfw/03_sample_asins.pkl")
